+++ terminal/input/example_test.go
@@ -7,7 +7,10 @@ package input
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-import "fmt"
+import (
+	"fmt"
+	"strings"
+)
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -16,20 +19,22 @@ func ExampleRead() {
 	input, err := Read("Please enter user name", true)
 
 	if err != nil {
-		fmt.Printf("Error: %v", err)
+		fmt.Printf("Error: %v\n", err)
+		return
 	}
 
-	fmt.Printf("User name: %s\v", input)
+	fmt.Printf("User name: %s\n", input)
 
 	// You can read user input without providing any title
 	fmt.Println("Please enter user name")
 	input, err = Read("", true)
 
 	if err != nil {
-		fmt.Printf("Error: %v", err)
+		fmt.Printf("Error: %v\n", err)
+		return
 	}
 
-	fmt.Printf("User name: %s\v", input)
+	fmt.Printf("User name: %s\n", input)
 }
 
 func ExampleReadPassword() {
@@ -41,10 +46,11 @@ func ExampleReadPassword() {
 	password, err := ReadPassword("Please enter password", true)
 
 	if err != nil {
-		fmt.Printf("Error: %v", err)
+		fmt.Printf("Error: %v\n", err)
+		return
 	}
 
-	fmt.Printf("User password: %s\v", password)
+	fmt.Printf("User password: %s\n", password)
 }
 
 func ExampleReadPasswordSecure() {
@@ -56,10 +62,11 @@ func ExampleReadPasswordSecure() {
 	password, err := ReadPasswordSecure("Please enter password", true)
 
 	if err != nil {
-		fmt.Printf("Error: %v", err)
+		fmt.Printf("Error: %v\n", err)
+		return
 	}
 
-	fmt.Printf("User password: %s\v", string(password.Data))
+	fmt.Printf("User password: %s\n", string(password.Data))
 
 	password.Destroy()
 }
@@ -76,4 +83,105 @@ func ExampleReadAnswer() {
 	if ok {
 		fmt.Println("File removed")
 	}
+}
+
+func ExampleAddHistory() {
+	input, err := Read("Please enter user name", true)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Save entered value to the input history
+	AddHistory(input)
+
+	fmt.Printf("User name: %s\n", input)
+}
+
+func ExampleSetHistoryCapacity() {
+	input, err := Read("Please enter user name", true)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Limit history size to last 3 entries
+	SetHistoryCapacity(3)
+
+	// Save entered value to the input history
+	AddHistory(input)
+
+	fmt.Printf("User name: %s\n", input)
+}
+
+func ExampleSetCompletionHandler() {
+	commands := []string{"add", "delete", "search", "help", "quit"}
+
+	input.SetCompletionHandler(func(input string) []string {
+		var result []string
+
+		for _, c := range commands {
+			if strings.HasPrefix(c, input) {
+				result = append(result, c)
+			}
+		}
+
+		return result
+	})
+
+	input.SetHintHandler(func(input string) string {
+		for _, c := range commands {
+			if strings.HasPrefix(c, input) {
+				return c[len(input):]
+			}
+		}
+
+		return ""
+	})
+
+	input, err := input.Read("Please enter command", true)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	fmt.Printf("Command: %s\n", input)
+}
+
+func ExampleSetHintHandler() {
+	commands := []string{"add", "delete", "search", "help", "quit"}
+
+	input.SetCompletionHandler(func(input string) []string {
+		var result []string
+
+		for _, c := range commands {
+			if strings.HasPrefix(c, input) {
+				result = append(result, c)
+			}
+		}
+
+		return result
+	})
+
+	input.SetHintHandler(func(input string) string {
+		for _, c := range commands {
+			if strings.HasPrefix(c, input) {
+				return c[len(input):]
+			}
+		}
+
+		return ""
+	})
+
+	input, err := input.Read("Please enter command", true)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	fmt.Printf("Command: %s\n", input)
 }