+++ terminal/input/input_windows.go
@@ -28,13 +28,11 @@ var MaskSymbol = "*"
 // ❗ HideLength is flag for hiding password length
 var HideLength = false
 
-var (
-	// ❗ MaskSymbolColorTag is fmtc color tag used for MaskSymbol output
-	MaskSymbolColorTag = ""
+// ❗ MaskSymbolColorTag is fmtc color tag used for MaskSymbol output
+var MaskSymbolColorTag = ""
 
-	// ❗ TitleColorTag is fmtc color tag used for input titles
-	TitleColorTag = "{s}"
-)
+// ❗ TitleColorTag is fmtc color tag used for input titles
+var TitleColorTag = "{s}"
 
 // ❗ HidePassword is flag for hiding password while typing
 // Because of using the low-level linenoise method for this feature, we can not use a
@@ -47,23 +45,23 @@ var AlwaysYes = false
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// ❗ Read reads user's input
+// ❗ Read reads user input
 func Read(title string, nonEmpty bool) (string, error) {
 	panic("UNSUPPORTED")
 }
 
-// ❗ ReadAnswer reads user's answer for yes/no question
+// ❗ ReadAnswer reads user's answer to yes/no question
 func ReadAnswer(title string, defaultAnswers ...string) (bool, error) {
 	panic("UNSUPPORTED")
 }
 
-// ❗ ReadPassword reads password or some private input which will be hidden
+// ❗ ReadPassword reads password or some private input that will be hidden
 // after pressing Enter
 func ReadPassword(title string, nonEmpty bool) (string, error) {
 	panic("UNSUPPORTED")
 }
 
-// ❗ ReadPasswordSecure reads password or some private input which will be hidden
+// ❗ ReadPasswordSecure reads password or some private input that will be hidden
 // after pressing Enter
 func ReadPasswordSecure(title string, nonEmpty bool) (*secstr.String, error) {
 	panic("UNSUPPORTED")
@@ -74,7 +72,7 @@ func AddHistory(data string) {
 	panic("UNSUPPORTED")
 }
 
-// ❗ SetCompletionHandler adds function for autocompletion
+// ❗ SetCompletionHandler adds autocompletion function (using Tab key)
 func SetCompletionHandler(h func(input string) []string) {
 	panic("UNSUPPORTED")
 }