+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Changelog
 
-### 12.117.1
+### 12.118.0
 
+- `[terminal/input]` Added method `SetHistoryCapacity`
 - `[options]` Improved `Errors.String` output formatting
+- `[terminal/input]` Added more usage examples
+- `[terminal/input]` Code refactoring
 
 ### 12.117.0